proyecto/requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
transbank-sdk==2.0.0
pandas==1.3.3  